route handler improvements

+++ src/app/api/cart/delete/by-id/route.ts
@@ -2,24 +2,41 @@ import { NextResponse } from 'next/server';
 
 import { CustomResponseType } from '@/types';
 import createSupabaseServerClient from '@/lib/supabase/supabase-server';
+import { noIdReceivedError, notAuthenticatedError } from '@/constants/api';
 
 export async function DELETE(request: Request): Promise<NextResponse<CustomResponseType>> {
+  const supabase = await createSupabaseServerClient();
+
   try {
-    const supabase = await createSupabaseServerClient();
+    const {
+      data: { session },
+    } = await supabase.auth.getSession();
 
     const cartItemId: string = await request.json();
+
+    if (!session)
+      return NextResponse.json({
+        success: false,
+        message: `Failed to remove item from cart. ${notAuthenticatedError}`,
+      });
+
+    if (!cartItemId)
+      return NextResponse.json({
+        success: false,
+        message: `Failed to remove item from cart. ${noIdReceivedError}`,
+      });
 
     const { error } = await supabase.from('cart').delete().eq('cartItemId', cartItemId);
 
     if (error) {
-      return NextResponse.json({ success: false, message: `Failed to remove product from cart. ${error.message}.` });
+      return NextResponse.json({ success: false, message: `Failed to remove item from cart. ${error.message}.` });
     }
 
-    return NextResponse.json({ success: true, message: 'Product remove from cart successfully.' });
+    return NextResponse.json({ success: true, message: 'Item remove from cart successfully.' });
   } catch (error) {
     return NextResponse.json({
       success: false,
-      message: 'Failed to remove product from cart. An unexpect error occured.',
+      message: 'Failed to remove item from cart. An unexpect error occured.',
     });
   }
 }
